+++ app/renderer/reducers/nodeSystem.reducer.js
@@ -70,7 +70,13 @@ export default function localStorage(state = DEFAULT_STATE, { type, payload }) {
     case types.nodeWsSystemChain.completed:
       return {
         ...state,
-        localNode: { ...state.localNode, chain: payload.result },
+        localNode: { ...state.localNode, chain: payload },
+      };
+
+    case types.nodeWsSystemChainRemote.completed:
+      return {
+        ...state,
+        remoteNode: { ...state.remoteNode, chain: payload },
       };
     default:
       return state;